Modify the profile manager and SBAUser to have one-way pointers
These changes are to allow accessing items that are stored and/or displayed by the profile manager while using the same method to access those items via the SBAUser (if defined for the profile manager). These changes are necessary to ensure that the user and profile manager objects are accessing the same values.