Move building task from SBBSurvey to background thread
There are a number of crashes when the survey factory is trying to create the survey, but the crash logs do not return an exception for *why* the app is crashing. Since we’ve received reports that the daily check-in is crashing the app, and that survey has a lot of customization, I am suspicious that the app is crashing b/c it was on the main thread so I’ve moved building the survey to the background queue.  Additionally, as a pair of suspenders, I am explicitly retaining the references. Swift should do this automagically, but figured it wouldn’t hurt. :)